improved hullmod scripts

+++ src/data/hullmods/sd_groundsupport.java
@@ -9,18 +9,23 @@ import com.fs.starfarer.api.ui.TooltipMakerAPI;
 import com.fs.starfarer.api.util.Misc;
 
 public class sd_groundsupport extends BaseHullMod {
-	final float GROUND_BONUS = 25;
+	final int GROUND_BONUS = 25;
 	public void applyEffectsBeforeShipCreation(HullSize hullSize, MutableShipStatsAPI stats, String id) {
 		stats.getDynamic().getMod(Stats.FLEET_GROUND_SUPPORT).modifyFlat(id, GROUND_BONUS);
 	}
 	@Override
 	public void addPostDescriptionSection(TooltipMakerAPI tooltip, ShipAPI.HullSize hullSize, ShipAPI ship, float width, boolean isForModSpec) {
-		tooltip.addPara("Increases the effective strength of planetary raids by "+ (int)GROUND_BONUS +", up to the total number of marines in the fleet.", 10f, Misc.getHighlightColor(), String.valueOf((int)GROUND_BONUS));
+		tooltip.addPara("Increases the effective strength of planetary raids by "+ GROUND_BONUS +", up to the total number of marines in the fleet.", 10f,
+				Misc.getHighlightColor(), String.valueOf(GROUND_BONUS));
 	}
 	@Override
-	public boolean shouldAddDescriptionToTooltip(ShipAPI.HullSize hullSize, ShipAPI ship, boolean isForModSpec) { return false; }
+	public boolean shouldAddDescriptionToTooltip(ShipAPI.HullSize hullSize, ShipAPI ship, boolean isForModSpec) {
+		return false;
+	}
 	@Override
-	public boolean showInRefitScreenModPickerFor(ShipAPI ship) { return ship.getHullSpec().getManufacturer().equals("???"); }
+	public boolean showInRefitScreenModPickerFor(ShipAPI ship) {
+		return ship.getHullSpec().getManufacturer().equals("???");
+	}
 }
